Move out app configuration
